pulling out the app host out to be handled by a new runtime-aware pipeline
